+++ changelog.md
@@ -3,6 +3,7 @@
 
 - Removed the "Show Attempts and Jumps on Platformer" feature since it already [EndscreenTweaks](mod:raydeeux.endscreentweaks) has it.
 - Added nullptr check for the background layer in the **Game Type Indicator** feature to prevent crashes.
+- Added a new feature called **Clickable Level Name to Browser**
 
 ## v1.0.0